Factor out docusign_views.py.
